Split event parsers into separate files, depending on format
