+++ sourcespec2/input/event_and_picks.py
@@ -18,8 +18,51 @@ Read event and phase picks.
 import sys
 import logging
 from ..setup import config, ssp_exit
-from .event_metadata import parse_qml, parse_hypo_file, parse_hypo71_picks
+from .event_parsers import (
+    parse_hypo71_hypocenter, parse_hypo71_picks, parse_hypo2000_file,
+    parse_qml_file, parse_source_spec_event_file
+)
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
+
+
+# pylint: disable=inconsistent-return-statements
+def _parse_hypo_file(hypo_file, event_id=None):
+    """
+    Parse a SourceSpec Event File, hypo71 or hypo2000 hypocenter file.
+
+    :param hypo_file: Path to the hypocenter file.
+    :type hypo_file: str
+    :param event_id: Event ID.
+    :type event_id: str
+
+    :return: A tuple of (SSPEvent, picks, format).
+    :rtype: tuple
+    """
+    err_msgs = []
+    parsers = {
+        'ssp_event_file': parse_source_spec_event_file,
+        'hypo71': parse_hypo71_hypocenter,
+        'hypo2000': parse_hypo2000_file,
+    }
+    format_strings = {
+        'ssp_event_file': 'SourceSpec Event File',
+        'hypo71': 'hypo71 hypocenter file',
+        'hypo2000': 'hypo2000 hypocenter file',
+    }
+    for file_format, parser in parsers.items():
+        try:
+            ssp_event, picks = parser(hypo_file, event_id)
+            return ssp_event, picks, file_format
+        except Exception as err:
+            format_str = format_strings[file_format]
+            msg = f'{hypo_file}: Not a {format_str}'
+            err_msgs.append(msg)
+            msg = f'Parsing error: {err}'
+            err_msgs.append(msg)
+    # If we arrive here, the file was not recognized as valid
+    for msg in err_msgs:
+        logger.error(msg)
+    ssp_exit(1)
 
 
 def _log_event_info(ssp_event):
@@ -50,7 +93,7 @@ def read_event_and_picks(trace1=None):
     ssp_event = None
     # parse hypocenter file
     if config.options.hypo_file is not None:
-        ssp_event, picks, file_format = parse_hypo_file(
+        ssp_event, picks, file_format = _parse_hypo_file(
             config.options.hypo_file, config.options.evid)
         config.hypo_file_format = file_format
     # parse pick file
@@ -58,7 +101,7 @@ def read_event_and_picks(trace1=None):
         picks = parse_hypo71_picks()
     # parse QML file
     if config.options.qml_file is not None:
-        ssp_event, picks = parse_qml()
+        ssp_event, picks = parse_qml_file()
     if ssp_event is not None:
         _log_event_info(ssp_event)
