+++ sourcespec2/input/event_parsers/obspy_catalog.py
@@ -1,0 +1,294 @@
+# -*- coding: utf-8 -*-
+# SPDX-License-Identifier: CECILL-2.1
+"""
+Parse event metadata and picks from an ObsPy catalog object.
+
+:copyright:
+    2012-2024 Claudio Satriano <satriano@ipgp.fr>
+:license:
+    CeCILL Free Software License Agreement v2.1
+    (http://www.cecill.info/licences.en.html)
+"""
+import contextlib
+import logging
+import re
+from ...setup import config, ssp_exit
+from ...ssp_event import SSPEvent
+from ...ssp_pick import SSPPick
+logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
+
+
+def _get_evid_from_resource_id(resource_id):
+    """
+    Get evid from resource_id.
+
+    :param resource_id: resource_id string
+    :type resource_id: str
+
+    :returns: evid string
+    :rtype: str
+    """
+    evid = resource_id
+    if '/' in evid:
+        evid = resource_id.split('/')[-1]
+    if '?' in evid:
+        evid = resource_id.split('?')[-1]
+    if '&' in evid:
+        evid = evid.split('&')[0]
+    if '=' in evid:
+        evid = evid.split('=')[-1]
+    return evid
+
+
+def _parse_event_metadata(obspy_event):
+    """
+    Parse event metadata from an ObsPy event object.
+
+    :param obspy_event: ObsPy event object
+    :type obspy_event: obspy.core.event.Event
+
+    :return: SSPEvent object
+    :rtype: SSPEvent
+    """
+    # No need to parse event name from QuakeML if event name is given
+    # in the command line
+    if config.options.evname is None:
+        parse_event_name_from_description = config.qml_event_description
+        event_description_regex = config.qml_event_description_regex
+    else:
+        parse_event_name_from_description = False
+        event_description_regex = None
+    ssp_event = SSPEvent()
+    ssp_event.event_id = _get_evid_from_resource_id(
+        str(obspy_event.resource_id.id))
+    if parse_event_name_from_description:
+        try:
+            ssp_event.name = str(obspy_event.event_descriptions[0].text)
+        except IndexError:
+            logger.warning(
+                'The event does not contain a description. Cannot parse '
+                'event name from description.'
+            )
+    if ssp_event.name and event_description_regex:
+        pattern = re.compile(event_description_regex)
+        match = pattern.search(ssp_event.name)
+        if match:
+            name = match.group()
+            # capitalize first letter
+            name = name[0].upper() + name[1:]
+            ssp_event.name = name
+    # See if there is a preferred origin...
+    origin = obspy_event.preferred_origin()
+    # ...or just use the first one
+    if origin is None:
+        origin = obspy_event.origins[0]
+    ssp_event.hypocenter.longitude.value_in_deg = origin.longitude
+    ssp_event.hypocenter.latitude.value_in_deg = origin.latitude
+    ssp_event.hypocenter.depth.value = origin.depth
+    ssp_event.hypocenter.depth.units = 'm'
+    ssp_event.hypocenter.origin_time = origin.time
+    return ssp_event
+
+
+def _parse_magnitude_from_obspy_event(obspy_event, ssp_event):
+    """
+    Parse magnitude from an ObsPy event.
+
+    :param obspy_event: ObsPy event
+    :type obspy_event: obspy.core.event.Event
+    :param ssp_event: SSPEvent object
+    :type ssp_event: SSPEvent
+    """
+    mag = obspy_event.preferred_magnitude() or obspy_event.magnitudes[0]
+    ssp_event.magnitude.value = mag.mag
+    ssp_event.magnitude.mag_type = mag.magnitude_type
+
+
+def _parse_scalar_moment_from_obspy_event(obspy_event, ssp_event):
+    """
+    Parse scalar moment from an ObsPy event.
+
+    :param obspy_event: ObsPy event
+    :type obspy_event: obspy.core.event.Event
+    :param ssp_event: SSPEvent object
+    :type ssp_event: SSPEvent
+    """
+    fm = obspy_event.preferred_focal_mechanism()\
+        or obspy_event.focal_mechanisms[0]
+    ssp_event.scalar_moment.value = fm.moment_tensor.scalar_moment
+    ssp_event.scalar_moment.units = 'N-m'
+
+
+def _parse_moment_tensor_from_obspy_event(obspy_event, ssp_event):
+    """
+    Parse moment tensor from an ObsPy event.
+
+    :param obspy_event: ObsPy event
+    :type obspy_event: obspy.core.event.Event
+    :param ssp_event: SSPEvent object
+    :type ssp_event: SSPEvent
+    """
+    fm = obspy_event.preferred_focal_mechanism()\
+        or obspy_event.focal_mechanisms[0]
+    mt = fm.moment_tensor.tensor
+    ssp_event.moment_tensor.m_rr = mt.m_rr
+    ssp_event.moment_tensor.m_tt = mt.m_tt
+    ssp_event.moment_tensor.m_pp = mt.m_pp
+    ssp_event.moment_tensor.m_rt = mt.m_rt
+    ssp_event.moment_tensor.m_rp = mt.m_rp
+    ssp_event.moment_tensor.m_tp = mt.m_tp
+    ssp_event.moment_tensor.units = 'N-m'
+
+
+def _parse_focal_mechanism_from_obspy_event(obspy_event, ssp_event):
+    """
+    Parse focal mechanism from an ObsPy event.
+
+    :param obspy_event: ObsPy event
+    :type obspy_event: obspy.core.event.Event
+    :param ssp_event: SSPEvent object
+    :type ssp_event: SSPEvent
+    """
+    fm = obspy_event.focal_mechanisms[0]
+    nodal_plane = fm.nodal_planes.nodal_plane_1
+    ssp_event.focal_mechanism.strike = nodal_plane.strike
+    ssp_event.focal_mechanism.dip = nodal_plane.dip
+    ssp_event.focal_mechanism.rake = nodal_plane.rake
+
+
+def _parse_picks_from_obspy_event(obspy_event):
+    """
+    Parse picks from an ObsPy event.
+
+    :param obspy_event: ObsPy event
+    :type obspy_event: obspy.core.event.Event
+
+    :return: list of SSPPick objects
+    :rtype: list
+    """
+    picks = []
+    for pck in obspy_event.picks:
+        pick = SSPPick()
+        pick.station = pck.waveform_id.station_code
+        pick.network = pck.waveform_id.network_code
+        pick.channel = pck.waveform_id.channel_code
+        if pck.waveform_id.location_code is not None:
+            pick.location = pck.waveform_id.location_code
+        else:
+            pick.location = ''
+        if pck.onset == 'emergent':
+            pick.flag = 'E'
+        elif pck.onset == 'impulsive':
+            pick.flag = 'I'
+        try:
+            pick.phase = pck.phase_hint[:1]
+        except Exception:
+            # ignore picks with no phase hint
+            continue
+        if pck.polarity == 'negative':
+            pick.polarity = 'D'
+        elif pck.polarity == 'positive':
+            pick.polarity = 'U'
+        pick.time = pck.time
+        picks.append(pick)
+    return picks
+
+
+def _get_event_from_obspy_catalog(
+        obspy_catalog, event_id=None, file_name=''):
+    """
+    Get an event from an ObsPy catalog object.
+
+    :param obspy_catalog: ObsPy catalog object
+    :type obspy_catalog: instance of :class:`obspy.core.event.Catalog`
+    :param event_id: event id
+    :type event_id: str
+    :param file_name: name of the file containing the catalog
+    :type file_name: str
+
+    :return: QuakeML event
+    :rtype: obspy.core.event.Event
+    """
+    if event_id is not None:
+        _obspy_events = [
+            ev for ev in obspy_catalog if event_id in str(ev.resource_id)
+        ]
+        try:
+            obspy_event = _obspy_events[0]
+        except IndexError as e:
+            raise ValueError(
+                f'Event {event_id} not found in {file_name}') from e
+    else:
+        obspy_event = obspy_catalog[0]
+        if len(obspy_catalog) > 1:
+            logger.warning(
+                f'Found {len(obspy_catalog)} events in {file_name}. '
+                'Using the first one.')
+    return obspy_event
+
+
+def _parse_obspy_event(obspy_event):
+    """
+    Parse event metadata and picks from an ObsPy event object.
+
+    :param obspy_event: ObsPy event object to parse
+    :type obspy_event: instance of :class:`obspy.core.event.Event`
+
+    :return: a tuple of (SSPEvent, picks)
+    :rtype: tuple
+    """
+    try:
+        ssp_event = _parse_event_metadata(obspy_event)
+        picks = _parse_picks_from_obspy_event(obspy_event)
+    except Exception as err:
+        logger.error(err)
+        ssp_exit(1)
+    log_messages = []
+    with contextlib.suppress(Exception):
+        _parse_moment_tensor_from_obspy_event(obspy_event, ssp_event)
+        # Compute focal mechanism, scalar moment and magnitude.
+        # They will be overwritten later on, if they are found in the
+        # QuakeML file.
+        ssp_event.focal_mechanism.from_moment_tensor(ssp_event.moment_tensor)
+        ssp_event.scalar_moment.from_moment_tensor(ssp_event.moment_tensor)
+        ssp_event.magnitude.from_scalar_moment(ssp_event.scalar_moment)
+    with contextlib.suppress(Exception):
+        _parse_scalar_moment_from_obspy_event(obspy_event, ssp_event)
+        # Compute magnitude from scalar moment. It will be overwritten later
+        # on, if it is found in the QuakeML file.
+        ssp_event.magnitude.from_scalar_moment(ssp_event.scalar_moment)
+    with contextlib.suppress(Exception):
+        _parse_magnitude_from_obspy_event(obspy_event, ssp_event)
+    with contextlib.suppress(Exception):
+        _parse_focal_mechanism_from_obspy_event(obspy_event, ssp_event)
+    for msg in log_messages:
+        logger.info(msg)
+    return ssp_event, picks
+
+
+def parse_obspy_catalog(obspy_catalog, event_id=None, file_name=''):
+    """
+    Parse event metadata and picks from an ObsPy catalog object.
+
+    :param obspy_catalog: ObsPy catalog object to parse
+    :type obspy_catalog: instance of :class:`obspy.core.event.Catalog`
+    :param event_id: event id to extract from the catalog. If None, the value
+        of config.options.evid is used. If this is also None,
+        the first event in the catalog is used.
+    :type event_id: str
+    :param file_name: name of the file containing the catalog
+    :type file_name: str
+
+    :return: a tuple of (SSPEvent, picks)
+    :rtype: tuple
+    """
+    event_id = event_id or config.options.evid
+    try:
+        obspy_event = _get_event_from_obspy_catalog(
+            obspy_catalog, event_id, file_name)
+    except Exception as err:
+        logger.error(err)
+        ssp_exit(1)
+    else:
+        ssp_event, picks = _parse_obspy_event(obspy_event)
+    return ssp_event, picks
